Some cleanup in the setup package
